Refactored bookmark related services + fix embed logic
