+++ projects/core/src/lib/pages/viewer-app/viewer-app.component.html
@@ -11,7 +11,7 @@
 </div>
 
 <div class="viewer" *ngIf="isLoaded" aria-label="Tailormap viewer page">
-  @if(isEmbedded) {
+  @if(isEmbedded$ | async) {
     <tm-embedded-layout></tm-embedded-layout>
   } @else {
     <tm-base-layout></tm-base-layout>
