+++ projects/core/src/lib/services/application-bookmark/bookmark.proto
