+++ projects/core/src/lib/services/application-bookmark/bookmark_pb.ts
