+++ projects/core/src/lib/services/application-bookmark/buf.gen.yaml
