+++ projects/core/src/lib/services/bookmark/bookmark.models.ts
